requirements.txt
python-dotenv
pydantic_settings